server/server-requirements.txt
pymongo
pyYAML
pika
flask
pyjwt
passlib
requests